Fixed bugs; almost working...
