+++ dbconn.js
@@ -1,12 +1,20 @@
 const dbconn = (function(){
-  const includes = require('./includes');
-  const mysql = includes.mysql,
-        dotenv = includes.dotenv.config();
+
+  const mysql = require('mysql'),
+        dotenv = require('dotenv').config();
   const connection = mysql.createConnection({
     host: process.env.DB_HOST,
     user: process.env.DB_USER,
     password: process.env.DB_PASSWORD,
     database: process.env.DB_NAME
+  });
+  connection.connect((err) => {
+    if (err) {
+      throw err;
+    }
+    else {
+      console.log(`Connected!`);
+    }
   });
   return connection;
 });
