+++ package.json
@@ -8,8 +8,8 @@
   },
   "dependencies": {
     "@sendgrid/mail": "^6.4.0",
-    "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
+    "bcrypt":"^3.0.6",
     "connect-flash": "^0.1.1",
     "cookie-parser": "^1.4.4",
     "debug": "~2.6.9",
@@ -17,6 +17,7 @@
     "ejs": "~2.6.1",
     "express": "^4.16.4",
     "express-session": "^1.16.2",
+    "express-ejs-layouts": "^2.5.0",
     "http-errors": "~1.6.3",
     "morgan": "~1.9.1",
     "multer": "^1.4.1",
